Add fix for AnyAsynchrononsUnitOfWork upstream not being called for some protocol function calls (#176)
## Description

The `AsynchrononsUnitOfWork` protocol has several functions that have default implementations that _can_ be implemented explicitly by workers (e.g. `cancel()`). However, the `AnyAsynchronousUnitOfWork` worker does not properly propagate these calls to the upstream, instead relying upon the default implementations, which results in these explicit implementations not being called.

One concrete example of this is the `Share` worker, which implements its own `cancel()` function. This is not properly being called when this worker is erased as an `AnyAsynchronousUnitOfWork` (see added test, which fails on the current production implementation).

In this PR, I've added all `AsynchronousUnitOfWork` protocol properties and functions explicitly to the `AnyAsynchronousUnitOfWork` type to utilize the upstream's implementation. This both fixes the current issue with the `Share` worker and ensures we avoid any future issues if a worker implements _any_ of the protocol properties or functions itself.